+++ modules/velociraptor-server/default.nix
@@ -1,0 +1,5 @@
+_: {
+  imports = [ ../../options/velociraptor ];
+
+  services.velociraptor.server.enable = true;
+}