+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - *(modules)* Add systemd-boot module - ([15d8754](https://github.com/JayRovacsek/nix-config/commit/15d875473a2d8eace6a7802bc737fdbece30009a))
 
 - *(modules,common,linux)* Resolve unifi not being nixified - ([6e6c9c2](https://github.com/JayRovacsek/nix-config/commit/6e6c9c2e439d4ad02c2c8f1e6325e3ee787d2ae9))
+
+- *(packages)* Add velociraptor package - ([bb4ba58](https://github.com/JayRovacsek/nix-config/commit/bb4ba58ad18e2b95e0b84d9128747f56bd363a65))
 - Update nix-options to 24.05 also - ([fab8c60](https://github.com/JayRovacsek/nix-config/commit/fab8c60c1629c0966a45665aff68ab4e208a4cf1))